+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of cookbook-openstack-net
 * Allow specifying the L3 agents' gateway_external_network by name.
 * Add nova_ca_certificates_file and nova_api_insecure; change nova_url to use the correct scheme
 * Fixed midonet and metaplugin plugin configuration file rendering
+* Make auth_version to be v2.0 in configuration file
 
 ## 10.0.1
 * Add tunnel_types item in ovs_neutron_plugin.ini.erb